pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "restaurant-rec"
version = "0.2.0"
description = "AI-powered restaurant recommendations (phase-wise architecture)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0.0",
  "datasets>=2.16.0",
  "huggingface_hub>=0.20.0",
  "groq>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]
